+++ gui/qtop.py
@@ -1,8 +1,8 @@
 from PySide.QtGui import QWidget, QPalette, QGridLayout, QPixmap, QLabel, QPushButton, QSizePolicy
 from PySide.QtCore import Qt
-from gui.chargui import char_creator
-from gui.pergui import per_creator
-from gui.slcontextgui import SL_creator
+from gui.char_creator.chargui import char_creator
+from gui.per_creator.pergui import per_creator
+from gui.sl_creator.slcontextgui import SL_creator
 from gui.supgui import sup_ui
 from libs import json_reader
