+++ gui/sl_creator/slinfo.py
