+++ gui/sl_creator/slview.py
